# openapi/calendar-app.openapi.json
{
  "components": {
    "schemas": {
      "AIConversationCreatePayload": {
        "additionalProperties": false,
        "properties": {
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "thread_id": {
            "anyOf": [
              {
                "maxLength": 64,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thread Id"
          },
          "title": {
            "default": "New AI conversation",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "title": "AIConversationCreatePayload",
        "type": "object"
      },
      "AIConversationMessagePayload": {
        "additionalProperties": false,
        "properties": {
          "content": {
            "maxLength": 20000,
            "minLength": 1,
            "title": "Content",
            "type": "string"
          },
          "message_id": {
            "anyOf": [
              {
                "maxLength": 64,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message Id"
          },
          "role": {
            "enum": [
              "user",
              "assistant"
            ],
            "title": "Role",
            "type": "string"
          },
          "structured": {
            "additionalProperties": true,
            "title": "Structured",
            "type": "object"
          }
        },
        "required": [
          "role",
          "content"
        ],
        "title": "AIConversationMessagePayload",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Calendar App API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/active-tool-journeys/funnel": {
      "get": {
        "operationId": "active_tool_funnel_baseline_api_active_tool_journeys_funnel_get",
        "parameters": [
          {
            "in": "query",
            "name": "template_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 80,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Template Id"
            }
          },
          {
            "in": "header",
            "name": "X-Desktop-Token",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Desktop-Token",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "calendar_session",
            "required": false,
            "schema": {
              "default": "",
              "title": "Calendar Session",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "calendar_csrf",
            "required": false,
            "schema": {
              "default": "",
              "title": "Calendar Csrf",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Active Tool Funnel Baseline Api Active Tool Journeys Funnel Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Active Tool Funnel Baseline"
      }
    },
    "/api/active-tool-journeys/{journey_id}/events": {
      "post": {
        "operationId": "record_active_tool_journey_event_api_active_tool_journeys__journey_id__events_post",
        "parameters": [
          {
            "in": "path",
            "name": "journey_id",
            "required": true,
            "schema": {
              "title": "Journey Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Desktop-Token",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Desktop-Token",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "calendar_session",
            "required": false,
            "schema": {
              "default": "",
              "title": "Calendar Session",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "calendar_csrf",
            "required": false,
            "schema": {
              "default": "",
              "title": "Calendar Csrf",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivationFunnelEventPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Record Active Tool Journey Event Api Active Tool Journeys  Journey Id  Events Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record Active Tool Journey Event"
      }
    },
...
    "/api/ai/conversations": {
      "get": {
        "operationId": "list_ai_conversations_api_ai_conversations_get",
        "parameters": [
          {
            "in": "query",
            "name": "include_archived",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Archived",
              "type": "boolean"
            }
          },
          {
            "in": "header",
            "name": "X-Desktop-Token",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Desktop-Token",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "calendar_session",
            "required": false,
            "schema": {
              "default": "",
              "title": "Calendar Session",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "calendar_csrf",
            "required": false,
            "schema": {
              "default": "",
              "title": "Calendar Csrf",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response List Ai Conversations Api Ai Conversations Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Ai Conversations"
      },
      "post": {
        "operationId": "create_ai_conversation_api_ai_conversations_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Desktop-Token",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Desktop-Token",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "calendar_session",
            "required": false,
            "schema": {
              "default": "",
              "title": "Calendar Session",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "calendar_csrf",
            "required": false,
            "schema": {
              "default": "",
              "title": "Calendar Csrf",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AIConversationCreatePayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Create Ai Conversation Api Ai Conversations Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Ai Conversation"
      }
    },
...
    "/api/ai/conversations/{thread_id}/messages": {
      "post": {
        "operationId": "add_ai_conversation_message_api_ai_conversations__thread_id__messages_post",
        "parameters": [
          {
            "in": "path",
            "name": "thread_id",
            "required": true,
            "schema": {
              "title": "Thread Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Desktop-Token",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Desktop-Token",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "calendar_session",
            "required": false,
            "schema": {
              "default": "",
              "title": "Calendar Session",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "calendar_csrf",
            "required": false,
            "schema": {
              "default": "",
              "title": "Calendar Csrf",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AIConversationMessagePayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Add Ai Conversation Message Api Ai Conversations  Thread Id  Messages Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Ai Conversation Message"
      }
    },
    "/api/ai/conversations/{thread_id}/messages/{message_id}/action-plan/dismiss": {
      "post": {
        "operationId": "dismiss_ai_action_plan_api_ai_conversations__thread_id__messages__message_id__action_plan_dismiss_post",
        "parameters": [
          {
            "in": "path",
            "name": "thread_id",
            "required": true,
            "schema": {
              "title": "Thread Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "message_id",
            "required": true,
            "schema": {
              "title": "Message Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Desktop-Token",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Desktop-Token",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "calendar_session",
            "required": false,
            "schema": {
              "default": "",
              "title": "Calendar Session",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "calendar_csrf",
            "required": false,
            "schema": {
              "default": "",
              "title": "Calendar Csrf",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Dismiss Ai Action Plan Api Ai Conversations  Thread Id  Messages  Message Id  Action Plan Dismiss Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Dismiss Ai Action Plan"
      }
    },
...
    "/api/check-ins/ensure": {
      "post": {
        "operationId": "ensure_check_ins_api_check_ins_ensure_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-Desktop-Token",
            "required": false,
            "schema": {
              "default": "",
              "title": "X-Desktop-Token",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "calendar_session",
            "required": false,
            "schema": {
              "default": "",
              "title": "Calendar Session",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "calendar_csrf",
            "required": false,
            "schema": {
              "default": "",
              "title": "Calendar Csrf",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Ensure Check Ins Api Check Ins Ensure Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ensure Check Ins"
      }
    },
...
  }
}